requirements.txt
python-dotenv>=1.0
pandas>=2.1
numpy>=1.26
requests>=2.31
APScheduler>=3.10
python-telegram-bot==22.5